+++ server/photomory/src/main/java/com/example/photomory/repository/AlbumRepository.java
@@ -5,6 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AlbumRepository extends JpaRepository<Album, Integer> {
-    // 기본 CRUD 기능은 JpaRepository가 이미 제공함
-}
+public interface AlbumRepository extends JpaRepository<Album, Long> { }
